+++ lib/srec/input/filter/and.cc
@@ -1,6 +1,6 @@
 //
 //      srecord - manipulate eprom load files
-//      Copyright (C) 1998, 1999, 2001, 2002, 2006-2008 Peter Miller
+//      Copyright (C) 1998, 1999, 2001, 2002, 2006-2009 Peter Miller
 //
 //      This program is free software; you can redistribute it and/or modify
 //      it under the terms of the GNU General Public License as published by
@@ -48,7 +48,7 @@ srec_input_filter_and::read(srec_record &record)
         return false;
     if (record.get_type() == srec_record::type_data)
     {
-        for (int j = 0; j < record.get_length(); ++j)
+        for (size_t j = 0; j < record.get_length(); ++j)
             record.set_data(j, record.get_data(j) & value);
     }
     return true;
