+++ lib/srec/output/file/brecord.cc
@@ -1,6 +1,6 @@
 //
 //      srecord - The "srecord" program.
-//      Copyright (C) 2007, 2008 Peter Miller
+//      Copyright (C) 2007-2009 Peter Miller
 //
 //      This program is free software; you can redistribute it and/or modify
 //      it under the terms of the GNU General Public License as published by
@@ -61,7 +61,7 @@ srec_output_file_brecord::write(const srec_record &record)
         break;
 
     case srec_record::type_data:
-        for (int j = 0; j < record.get_length(); ++j)
+        for (size_t j = 0; j < record.get_length(); ++j)
             buffer_stash(record.get_address() + j, record.get_data(j));
         break;
     }
